+++ tests/testthat/test-prism_archive_clean.R
@@ -1,4 +1,4 @@
 
-test_that("prism_archive_clean() works", {
-  expect_true(TRUE)
+test_that("prism_archive_clean() erros", {
+  expect_error(prism_archive_clean("anything"))
 })